Moved createSafeContext util inside common.utils.ts
